Add files via upload
board.h과 board.cpp로 분리해서 구현 완료.
